Renamed font-awesone-helper.js to font-awesome-helper.js
