+++ src/static/styles/font-awesome-helper.js
