+++ src/static/styles/font-awesome.config.prod.js
@@ -1,6 +1,6 @@
 const ExtractTextPlugin = require('extract-text-webpack-plugin');
 const fontAwesomeConfig = require('./font-awesome.config');
-const buildExtractStylesLoader = require('./font-awesone-helper');
+const buildExtractStylesLoader = require('./font-awesome-helper');
 
 fontAwesomeConfig.styleLoader = buildExtractStylesLoader(ExtractTextPlugin.extract(['css-loader', 'less-loader']));
